Merged PR 43056: Remove unnecessary call to ReleaseStr
When `RemoveUpgradeRelatedBundle` is set to an invalid value, it should be set back to `always`. It's not necessary to call `ReleaseStr` because `StrAllocString` will try to reuse the previously allocated buffer and shrink/expand it as necessary. By calling `ReleaseStr`, the memory is released, but the pointer is not `NULL` resulting in a crash.